稼働時間の表示

+++ src/main/java/jp/co/waja/app/util/WorkTimes.java
@@ -1,10 +1,11 @@
 package jp.co.waja.app.util;
 
+import java.math.BigDecimal;
 import java.time.LocalTime;
 
 public class WorkTimes {
 
-	public Long workTime(LocalTime startAt, LocalTime endAt, Integer restTime) {
+	public BigDecimal workTime(LocalTime startAt, LocalTime endAt, Integer restTime) {
 		return WorkTimeUtils.workTime(startAt, endAt, restTime);
 	}
 }
